convert all Int type of ID to Long/String in seller_home module

+++ features/merchant/seller_menu_common/src/main/java/com/tokopedia/seller/menu/common/domain/usecase/GetShopBadgeUseCase.kt
@@ -18,7 +18,7 @@ class GetShopBadgeUseCase @Inject constructor(private val gqlRepository: Graphql
 
         private const val SHOP_ID_KEY = "shopIds"
 
-        fun createRequestParams(shopId: Int) = HashMap<String, Any>().apply {
+        fun createRequestParams(shopId: Long) = HashMap<String, Any>().apply {
             put(SHOP_ID_KEY, listOf(shopId))
         }
     }
